Fix: Circular Dependency solved
