+++ libs/common/src/services/redis/redis.service.module.ts
@@ -1,10 +1,8 @@
 import { Module } from '@nestjs/common';
 
-import { RedisModule } from './redis.module';
 import { RedisService } from './redis.service';
 
 @Module({
-  imports: [RedisModule],
   providers: [RedisService],
   exports: [RedisService],
 })
