+++ src/controllers/cart/cart.controller.spec.ts
@@ -2,14 +2,10 @@ import { CommandBus } from '@nestjs/cqrs';
 import { Test, TestingModule } from '@nestjs/testing';
 import { mockDeep } from 'jest-mock-extended';
 
-import { productDetailsDtoMock } from '@mp/common/testing';
-
 import { CartController } from './cart.controller';
-import { SaveProductRedisCommand } from './command/save-product-redis.command';
-
 describe('CartController', () => {
   let controller: CartController;
-  let commandBus: CommandBus;
+  // let commandBus: CommandBus;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
@@ -23,26 +19,10 @@ describe('CartController', () => {
     }).compile();
 
     controller = module.get<CartController>(CartController);
-    commandBus = module.get<CommandBus>(CommandBus);
+    // commandBus = module.get<CommandBus>(CommandBus);
   });
 
   it('should be defined', () => {
     expect(controller).toBeDefined();
-  });
-
-  describe('SaveProductToRedis', () => {
-    it('should call execute on the commandBus with correct parameters', async () => {
-      // Arrange
-      const executeSpy = jest.spyOn(commandBus, 'execute');
-      const expectedCommand = new SaveProductRedisCommand(
-        productDetailsDtoMock.id,
-      );
-
-      // Act
-      await controller.saveProductToRedis(productDetailsDtoMock.id);
-
-      // Assert
-      expect(executeSpy).toHaveBeenCalledWith(expectedCommand);
-    });
   });
 });
