+++ src/controllers/product/product.module.ts
@@ -3,6 +3,7 @@ import { Module } from '@nestjs/common';
 import { CartServiceModule } from './../../domain/service/cart/cart.service.module';
 import { CreateProductCommandHandler } from './command/create-product.command.handler';
 import { DeleteProductCommandHandler } from './command/delete-product.command.handler';
+import { SaveProductRedisCommand } from './command/save-product-redis.command';
 import { SearchProductQueryHandler } from './command/search-product-query.handler';
 import { UpdateEnabledProductCommandHandler } from './command/update-enabled-product.command.handler';
 import { UpdateProductCommandHandler } from './command/update-product.command.handler';
@@ -16,6 +17,7 @@ const commandHandlers = [
   UpdateProductCommandHandler,
   UpdateEnabledProductCommandHandler,
   DeleteProductCommandHandler,
+  SaveProductRedisCommand,
 ];
 
 @Module({
